--- LaserChessGamePlay/background/Background.java
@@ -1,13 +1,0 @@
-package background;
-
-public class Background {
-    public enum Color {
-        RED, BLUE
-    }
-
-    public Color color;
-
-    public Background(Color color) {
-        this.color = color;
-    }
-}
