--- LaserChessGamePlay/chess/Chess.java
@@ -1,10 +1,0 @@
-package chess;
-
-public class Chess {
-    public enum Color {
-        RED, BLUE
-    }
-    public Color color;
-
-    public void rotate(char d){}//pass
-}
