--- LaserChessGamePlay/chess/ChessKing.java
@@ -1,6 +1,0 @@
-package chess;
-public class ChessKing extends Chess {
-    public ChessKing(Color color) {
-        this.color = color;
-    }
-}
