--- LaserChessGamePlay/chess/ChessTwoWayMirror.java
@@ -1,22 +1,0 @@
-package chess;
-public class ChessTwoWayMirror extends Chess {
-    public enum Direction {
-        LEFT_TOP, RIGHT_TOP
-    }
-
-    public Direction direction;
-
-    public ChessTwoWayMirror(Direction direction, Color color) {
-        this.direction = direction;
-        this.color = color;
-    }
-
-    public void rotate(char d){
-        if(direction==Direction.LEFT_TOP){
-            direction = Direction.RIGHT_TOP;
-        }
-        else{
-            direction = Direction.LEFT_TOP;
-        }
-    }
-}
