--- LaserChessGamePlay/operate/Operate.java
@@ -1,7 +1,0 @@
-package operate;
-
-public class Operate {
-    public void print() {
-        System.out.println("Generic Operation");
-    }
-}
