+++ src/main/java/com/example/flcw/LaserChessGamePlay/background/Background.java
@@ -1,0 +1,13 @@
+package com.example.flcw.LaserChessGamePlay.background;
+
+public class Background {
+    public enum Color {
+        RED, BLUE
+    }
+
+    public Color color;
+
+    public Background(Color color) {
+        this.color = color;
+    }
+}
