+++ src/main/java/com/example/flcw/LaserChessGamePlay/chess/Chess.java
@@ -1,0 +1,10 @@
+package com.example.flcw.LaserChessGamePlay.chess;
+
+public class Chess {
+    public enum Color {
+        RED, BLUE
+    }
+    public Color color;
+
+    public void rotate(char d){}//pass
+}
