+++ src/main/java/com/example/flcw/LaserChessGamePlay/chess/ChessKing.java
@@ -1,0 +1,6 @@
+package com.example.flcw.LaserChessGamePlay.chess;
+public class ChessKing extends Chess {
+    public ChessKing(Color color) {
+        this.color = color;
+    }
+}
