+++ src/main/java/com/example/flcw/LaserChessGamePlay/operate/Operate.java
@@ -1,0 +1,7 @@
+package com.example.flcw.LaserChessGamePlay.operate;
+
+public class Operate {
+    public void print() {
+        System.out.println("Generic Operation");
+    }
+}
