+++ src/main/java/module-info.java
@@ -1,0 +1,13 @@
+module com.example.flcw {
+    requires javafx.controls;
+    requires javafx.fxml;
+
+    requires org.controlsfx.controls;
+    requires com.dlsc.formsfx;
+    requires org.kordamp.ikonli.javafx;
+    requires org.kordamp.bootstrapfx.core;
+    requires com.almasb.fxgl.all;
+
+    opens com.example.flcw to javafx.fxml;
+    exports com.example.flcw;
+}
